Move the ArroyReader into its module
